Updated execSQL method

+++ NodeDevelopment/Index.js
@@ -21,35 +21,53 @@ app.use("/", router);
 app.listen(port);
 console.log("O servidor está ativo!");
 
-function execSQL(conn, sqlQry, res) {
-  conn
-    .request()
-    .query(sqlQry)
-    .then((result) => console.log("Sucesso " + result))
-    .catch((err) => console.log("Erro: " + err));
+async function execSQL(conn, sqlQry) {
+  return await conn.request().query(sqlQry);
 }
 
-router.post("/resultados", (req, res) => {
+router.post("/resultados", async (req, res) => {
   const RA = parseInt(req.body.ra);
   const Cod = parseInt(req.body.cod);
   const Nota = parseFloat(req.body.nota);
   const Frequencia = parseFloat(req.body.freq);
 
-  execSQL(
+  const existsAluno = await execSQL(
     global.conn,
-    "INSERT INTO RESULTADOS(RA,COD,NOTA,FREQUENCIA) VALUES(" +
-      RA +
-      "," +
-      Cod +
-      "," +
-      Nota +
-      "," +
-      Frequencia +
-      ")",
-    res
+    "SELECT * FROM ALUNOS WHERE RA =" + RA
   );
 
-  execSQL(global.conn, "DELETE FROM MATRICULAS WHERE RA=" + RA);
+  const existsDisciplina = await execSQL(
+    global.conn,
+    "SELECT * FROM DISCIPLINAS WHERE COD=" + Cod
+  );
+  const existsMatricula = await execSQL(
+    global.conn,
+    "SELECT * FROM MATRICULAS WHERE RA=" + RA
+  );
 
-  return res.json();
+  if (existsAluno.recordset.length == 0) {
+    return res.json(404);
+  } else if (existsDisciplina.recordset.length == 0) {
+    return res.json(404);
+  } else if (existsMatricula.recordset.length == 0) {
+    return res.json(404);
+  } else {
+    execSQL(
+      global.conn,
+      "INSERT INTO RESULTADOS(RA,COD,NOTA,FREQUENCIA) VALUES(" +
+        RA +
+        "," +
+        Cod +
+        "," +
+        Nota +
+        "," +
+        Frequencia +
+        ")",
+      res
+    );
+
+    execSQL(global.conn, "DELETE FROM MATRICULAS WHERE RA=" + RA);
+
+    return res.json(200);
+  }
 });
